refactor: move routing from kitsu/auth to auth/kitsu
Move to auth-kitsu so that all the auth packages can be put together, instead of the different
services put together.
