+++ src/main/kotlin/com/chesire/orochi/routes/Routing.kt
@@ -1,6 +1,6 @@
 package com.chesire.orochi.routes
 
-import com.chesire.orochi.routes.kitsu.kitsu
+import com.chesire.orochi.routes.auth.auth
 import com.chesire.orochi.routes.status.status
 import io.ktor.application.Application
 import io.ktor.routing.routing
@@ -11,6 +11,6 @@ import io.ktor.routing.routing
 fun Application.startRouting() {
     routing {
         status()
-        kitsu()
+        auth()
     }
 }
