+++ src/main/kotlin/com/chesire/orochi/routes/auth/kitsu/KitsuAuthService.kt
@@ -1,4 +1,4 @@
-package com.chesire.orochi.routes.kitsu.auth
+package com.chesire.orochi.routes.auth.kitsu
 
 import com.chesire.orochi.api.kitsu.KitsuEndpoint
 import com.chesire.orochi.api.kitsu.model.KitsuAuthFailureDto
